% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camDogs.R
\docType{data}
\name{camDogs}
\alias{camDogs}
\title{Dogs of Cambridge, MA}
\format{
A tibble with 3,942 rows and 7 variables:
\describe{
\item{Dog_Name}{a character vector of dog names}
}
}
\source{
{Fill in!}
}
\usage{
camDogs
}
\description{
Dogs of Cambridge, MA
}
\keyword{datasets}
