% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top10.R
\name{top10}
\alias{top10}
\title{Title}
\usage{
top10(data, x)
}
\arguments{
\item{data}{A data frame}

\item{x}{A character or factor vector used to determine most common}
}
\value{
A filtered data frame that only includes rows for the top 10 most common x values
}
\description{
Title
}
